# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API Human Link",
    "version": "3.0.0",
    "description": "Généré depuis les routeurs Express et les schémas zod partagés (shared/src/index.ts). Authentification par cookie httpOnly (`hl_at`). Les écritures (POST, PUT, PATCH, DELETE) exigent aussi l’en-tête `X-CSRF-Token` recopiant le cookie `hl_csrf` (obtenu par GET /api/auth/csrf). Les dates sont au format AAAA-MM-JJ."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/documents/user/{id}": {
      "get": {
        "tags": [
          "documents"
        ],
        "operationId": "get_documents_user_id",
        "summary": "Coffre-fort d’un employé",
        "description": "Consultation tracée dans le journal d’audit. Rôles autorisés : admin. Authentification requise (cookie de session).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "category": {
                            "type": "string",
                            "enum": [
                              "bulletin",
                              "contrat",
                              "autre"
                            ]
                          },
                          "label": {
                            "type": "string"
                          },
                          "month": {
                            "type": "number"
                          },
                          "year": {
                            "type": "number"
                          },
                          "size": {
                            "type": "number"
                          },
                          "createdAt": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "category",
                          "label",
                          "size",
                          "createdAt"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "items"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié ou session expirée"
          },
          "403": {
            "description": "Rôle insuffisant"
          },
          "default": {
            "description": "Erreur",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "x-documented": true,
        "x-roles": [
          "admin"
        ]
      },
      "post": {
        "tags": [
          "documents"
        ],
        "operationId": "post_documents_user_id",
        "summary": "Déposer un document dans le coffre d’un employé",
        "description": "PDF, JPEG ou PNG, 10 Mo au plus ; chiffré au repos et analysé par l’antivirus s’il est configuré. Rôles autorisés : admin. Authentification requise (cookie de session).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "bulletin",
                      "contrat",
                      "autre"
                    ]
                  },
                  "month": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 12
                  },
                  "year": {
                    "type": "integer",
                    "minimum": 1990,
                    "maximum": 2100
                  },
                  "name": {
                    "type": "string",
                    "maxLength": 120
                  },
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "type"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "file": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id"
                      ],
                      "additionalProperties": {}
                    }
                  },
                  "required": [
                    "file"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié ou session expirée"
          },
          "403": {
            "description": "Rôle insuffisant"
          },
          "default": {
            "description": "Erreur",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "x-documented": true,
        "x-roles": [
          "admin"
        ]
      }
    },
...
    "/api/users/{id}/anonymize": {
      "post": {
        "tags": [
          "users"
        ],
        "operationId": "post_users_id_anonymize",
        "summary": "Anonymiser un compte archivé (RGPD)",
        "description": "Irréversible : données personnelles, fichiers et notifications effacés. Rôles autorisés : admin. Authentification requise (cookie de session).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "ok"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié ou session expirée"
          },
          "403": {
            "description": "Rôle insuffisant"
          },
          "default": {
            "description": "Erreur",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "x-documented": true,
        "x-roles": [
          "admin"
        ]
      }
    },
    "/api/users/{id}/photo": {
      "get": {
        "tags": [
          "users"
        ],
        "operationId": "get_users_id_photo",
        "summary": "Photo de profil",
        "description": "Authentification requise (cookie de session).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fichier",
            "content": {
              "image/*": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié ou session expirée"
          },
          "default": {
            "description": "Erreur",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "x-documented": true
      }
    },
    "/api/users/{id}/resend-invite": {
      "post": {
        "tags": [
          "users"
        ],
        "operationId": "post_users_id_resend_invite",
        "summary": "Renvoyer l’invitation",
        "description": "Rôles autorisés : admin. Authentification requise (cookie de session).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "ok"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié ou session expirée"
          },
          "403": {
            "description": "Rôle insuffisant"
          },
          "default": {
            "description": "Erreur",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "x-documented": true,
        "x-roles": [
          "admin"
        ]
      }
    },
    "/api/users/{id}/reset-2fa": {
      "post": {
        "tags": [
          "users"
        ],
        "operationId": "post_users_id_reset_2fa",
        "summary": "Réinitialiser la 2FA d’un compte",
        "description": "Rôles autorisés : admin. Authentification requise (cookie de session).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "ok"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié ou session expirée"
          },
          "403": {
            "description": "Rôle insuffisant"
          },
          "default": {
            "description": "Erreur",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "x-documented": true,
        "x-roles": [
          "admin"
        ]
      }
    },
    "/api/users/{id}/restore": {
      "post": {
        "tags": [
          "users"
        ],
        "operationId": "post_users_id_restore",
        "summary": "Restaurer un compte archivé",
        "description": "Rôles autorisés : admin. Authentification requise (cookie de session).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "matricule": {
                          "type": "string"
                        },
                        "nom": {
                          "type": "string"
                        },
                        "prenom": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "role": {
                          "type": "string",
                          "enum": [
                            "admin",
                            "manager",
                            "employe"
                          ]
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "invited",
                            "active",
                            "archived",
                            "anonymized"
                          ]
                        },
                        "isDirector": {
                          "type": "boolean"
                        },
                        "poste": {
                          "type": "string"
                        },
                        "departement": {
                          "type": "string"
                        },
                        "sexe": {
                          "type": "string",
                          "enum": [
                            "Homme",
                            "Femme"
                          ]
                        },
                        "dateNaissance": {
                          "type": "string"
                        },
                        "nationalite": {
                          "type": "string"
                        },
                        "pays": {
                          "type": "string"
                        },
                        "ville": {
                          "type": "string"
                        },
                        "codePostal": {
                          "type": "string"
                        },
                        "adresse": {
                          "type": "string"
                        },
                        "telephone": {
                          "type": "string"
                        },
                        "numeroSecu": {
                          "type": "string"
                        },
                        "salaire": {
                          "type": "number"
                        },
                        "dateEmbauche": {
                          "type": "string"
                        },
                        "typeContrat": {
                          "type": "string",
                          "enum": [
                            "CDI",
                            "CDD",
                            "Alternance",
                            "Stage",
                            "Freelance"
                          ]
                        },
                        "soldeConge": {
                          "type": "number"
                        },
                        "teleworkMax": {
                          "type": "number"
                        },
                        "photoFileId": {
                          "type": "string"
                        },
                        "twoFactor": {
                          "type": "object",
                          "properties": {
                            "enabled": {
                              "type": "boolean"
                            }
                          },
                          "required": [
                            "enabled"
                          ],
                          "additionalProperties": false
                        }
                      },
                      "required": [
                        "id",
                        "matricule",
                        "nom",
                        "prenom",
                        "email",
                        "role",
                        "status",
                        "isDirector",
                        "soldeConge",
                        "teleworkMax"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "user"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié ou session expirée"
          },
          "403": {
            "description": "Rôle insuffisant"
          },
          "default": {
            "description": "Erreur",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "x-documented": true,
        "x-roles": [
          "admin"
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "hl_at"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {}
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      }
    }
  }
}